EnvVarEcho_To_CSVandText
This script just makes it easier for me to see the Path variable's list of paths from command prompt's ridiculous output.\nThe use of it is easy. Just have cmd prompt pipe its output of the command echo %PATH% into some text file and then run this script with that text file as input (name the file in a command line argument).