Fixed slight bug in printing results. Using regex instead of fnfmatch now.
Bug was that  was actually a list of lists instead of a flat list like I initially thought. So need to iterate through the sublists. This was noticed because I noticed that only one result from a given directory in PATH was ever printed even though I knew there was definitely more. I also opted for regex instead of fnfmatch-style searching because I am used to using grep and think that would give me more power than UNIX find kind of search.